fix custom scale factor parsing or losing precision on float -> double cast

+++ src/main/java/org/xtx/ut4converter/MapConverter.java
@@ -1337,7 +1337,7 @@ public class MapConverter extends Task<T3DLevelConvertor> {
 		return t3dLvlConvertor;
 	}
 
-	public void setScale(float scale) {
+	public void setScale(Double scale) {
 		this.scale = scale;
 	}
